Hook up and use the new hough finder code.

+++ larpandoradlcontent/LArSlicing/DlSlicingAlgorithm.h
@@ -66,7 +66,7 @@ private:
     LArDLHelper::TorchModel m_modelFile; ///< The model to use.
 
     float m_scalingFactor;            ///< The scaling factor for the input.
-    std::vector<double> m_thresholds; ///< Distance Class Thresholds.
+    std::vector<float> m_thresholds;  ///< Distance Class Thresholds.
     int m_nDistanceClasses;           ///< The number of distance classes (derived from thresholds).
     std::string m_caloHitListName;    ///< The name of the input CaloHit list.
     int m_k;                          ///< The number of nearest neighbours to use when building the graph.
